broke up parallel into package
