+++ autofit/non_linear/parallel/__init__.py
@@ -1,0 +1,5 @@
+from .process import AbstractJob
+from .process import AbstractJobResult
+from .process import Process
+from .sneaky import SneakyJob
+from .sneaky import SneakyPool
